+++ internal/use_cases/post_expression/type.go
@@ -1,7 +1,7 @@
 package postexpression
 
 type request struct {
-	userId     int64
+	userId     int
 	ExitId     string `json:"exit_id"`
 	Expression string `json:"expression"`
 }
